pyproject.toml
[project]
name = "embedding"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "branca>=0.8.2",
    "matplotlib>=3.10.8",
    "pyvis>=0.3.2",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.2.2",
    "umap-learn>=0.5.11",
]
